undo bad renames
